removed ssl and removed allocations
